64-bit patch

+++ challenges/netstorage/lib/cgc_malloc.h
@@ -38,7 +38,7 @@ extern cgc_size_t size_class_limits[NUM_FREE_LISTS];
 
 struct blk_t {
   cgc_size_t size;
-  unsigned int free;
+  unsigned long free;
   struct blk_t *fsucc;
   struct blk_t *fpred;
   struct blk_t *next;
